add notify action
