+++ aether/actions/action.h
@@ -161,71 +161,6 @@ class Action : public IAction {
   ActionTrigger* action_trigger_{};
   ActionRegistry::IndexShare index_;
 };
-
-template <typename T = void>
-class NotifyAction : public Action<NotifyAction<T>> {
- public:
-  using Action<NotifyAction<T>>::Action;
-  using Action<NotifyAction<T>>::operator=;
-
-  TimePoint Update(TimePoint current_time) override {
-    if (notify_success_) {
-      notify_success_ = false;
-      this->ResultRepeat(static_cast<T&>(*this));
-    }
-    if (notify_failed_) {
-      notify_failed_ = false;
-      this->Error(static_cast<T&>(*this));
-    }
-    return current_time;
-  }
-
-  void Notify() {
-    notify_success_ = true;
-    this->Trigger();
-  }
-  void Failed() {
-    notify_failed_ = true;
-    this->Trigger();
-  }
-
- private:
-  bool notify_success_{};
-  bool notify_failed_{};
-};
-
-template <>
-class NotifyAction<void> : public Action<NotifyAction<void>> {
- public:
-  using Action<NotifyAction<void>>::Action;
-  using Action<NotifyAction<void>>::operator=;
-
-  TimePoint Update(TimePoint current_time) override {
-    if (notify_success_) {
-      notify_success_ = false;
-      this->ResultRepeat(*this);
-    }
-    if (notify_failed_) {
-      notify_failed_ = false;
-      this->Error(*this);
-    }
-    return current_time;
-  }
-
-  void Notify() {
-    notify_success_ = true;
-    this->Trigger();
-  }
-  void Failed() {
-    notify_failed_ = true;
-    this->Trigger();
-  }
-
- private:
-  bool notify_success_{};
-  bool notify_failed_{};
-};
-
 }  // namespace ae
 
 #endif  // AETHER_ACTIONS_ACTION_H_
